xen: separate the hypervisor recipe from the tools and uprev to 4.12.1
This enables efficient building of the hypervisor and tools separately
eg. for different architectures.

xen-image-minimal updated: the new xen-tools package replaces xen-base.

Uprev the SRCREV in the git recipe to a newer revision that builds
successfully with the current compiler, which requires adding packaging
for the new xen-ucode tool, introduced in Xen 4.13 development cycle for
late microcode update.

OpenXT ticket reference: OXT-1694

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>