xen: pass compiler flags via Xen-provided variables
Fix the method of passing compiler flags into the Xen build,
needed to fix hvmloader: it must not be compiled with SSE options
which are provided via TUNE_CCARGS. hvmloader is needed for Xen
to be able to run HVM-mode guest VMs.

32-bit ARM needs the TUNE_CCARGS to be retained as passed via CC
to ensure that configure is able to generate binaries for the
correct ABI, so that target platform is handled as a separate case.

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>