REVISIT ME: On ARM platforms, include the hypervisor in the dom0 filesystem
Xen shouldn't need to be in any domain filesystem, but the FATPAYLOAD
method of adding files to the boot filesystem image of the Raspberry Pi
SD card image does pull files from there, so this is a workaround to
enable using that.
Instead, there needs to be an alternatve way to add files, such as
the hypervisor boot binary/binaries, to the SD card boot partition.